+++ src/services/Login.php
@@ -9,8 +9,13 @@
 namespace flipbox\saml\sp\services;
 
 use craft\base\Component;
+use craft\elements\User;
 use flipbox\saml\core\exceptions\InvalidMessage;
+use flipbox\saml\core\helpers\MessageHelper;
 use flipbox\saml\sp\events\UserLogin;
+use flipbox\saml\sp\models\Settings;
+use flipbox\saml\sp\records\ProviderIdentityRecord;
+use flipbox\saml\sp\records\ProviderRecord;
 use flipbox\saml\sp\Saml;
 use flipbox\saml\sp\services\login\AssertionTrait;
 use SAML2\Response as SamlResponse;
@@ -27,24 +32,13 @@ class Login extends Component
     const EVENT_BEFORE_RESPONSE_TO_USER = 'eventBeforeResponseToUser';
     const EVENT_AFTER_RESPONSE_TO_USER = 'eventAfterResponseToUser';
 
-    /**
-     * @param SamlResponse $response
-     * @return \flipbox\saml\sp\records\ProviderIdentityRecord
-     * @throws InvalidMessage
-     * @throws UserException
-     * @throws \Exception
-     * @throws \Throwable
-     * @throws \craft\errors\ElementNotFoundException
-     * @throws \yii\base\Exception
-     */
-    public function login(
-        SamlResponse $response
+    public function transformToUser(
+        User $user,
+        SamlResponse $response,
+        ProviderRecord $idp,
+        ProviderRecord $sp,
+        Settings $settings
     ) {
-        /**
-         * Get User
-         */
-        $user = Saml::getInstance()->getUser()->getByResponse($response);
-
         /**
          * Before user save
          */
@@ -56,44 +50,51 @@ class Login extends Component
             static::EVENT_BEFORE_RESPONSE_TO_USER,
             $event
         );
-        /**
-         * Sync User
-         */
-        Saml::getInstance()->getUser()->sync($user, $response);
+
+        // Sync
+        Saml::getInstance()->getUser()->sync(
+            $user,
+            $response,
+            $idp,
+            $sp,
+            $settings
+        );
 
         /**
-         * after user save
+         * After user save
          */
         $event = new UserLogin();
         $event->response = $response;
         $event->user = $user;
 
         $this->trigger(
-            static::EVENT_AFTER_RESPONSE_TO_USER,
+            static::EVENT_BEFORE_RESPONSE_TO_USER,
             $event
         );
 
-        /**
-         * Get Identity
-         */
-        $identity = Saml::getInstance()->getProviderIdentity()->getByUserAndResponse($user, $response);
+        return $user;
+    }
 
+    /**
+     * @param ProviderIdentityRecord $identityRecord
+     * @throws UserException
+     * @throws \Throwable
+     */
+    public function byIdentity(ProviderIdentityRecord $identityRecord)
+    {
         /**
          * Log user in
          */
-        if (! Saml::getInstance()->getUser()->login($identity)) {
+        if (! Saml::getInstance()->getUser()->login($identityRecord)) {
             throw new UserException("Unknown error while logging in.");
         }
-
         /**
          * User's successfully logged in so we can now set the lastLogin for the
          * provider identity and save it to the db.
          */
-        $identity->lastLoginDate = new \DateTime();
-        if (! Saml::getInstance()->getProviderIdentity()->save($identity)) {
+        $identityRecord->lastLoginDate = new \DateTime();
+        if (! Saml::getInstance()->getProviderIdentity()->save($identityRecord)) {
             throw new UserException("Error while saving identity.");
         }
-
-        return $identity;
     }
 }